+++ massiv/CHANGELOG.md
@@ -1,3 +1,12 @@
+# 0.2.4
+
+* Addition of inner folds: `ifoldlInner`, `foldlInner`, `ifoldrInner` and `foldrInner`
+* Addition of functions that can fold over any dimension (`foldlWithin`, `foldlWithin', etc.)
+* Addition of `ifoldMono` and `ifoldSemi`, thus fixing: [#
+* Improvement over manipulating index dimensions with addition of type level `Dimension n` data type
+  and functions like `getDimension`, `dropDimension`.
+* Addition of `insertDim` and type level `insertDimension`
+
 # 0.2.3
 
 * Addition of `Profunctor` functions for `Stencil`: `lmapStencil`, `rmapStencil` and `bimapStencil`